Conflicts:
	src/dl_txt_unpack.cpp

+++ src/dl_convert.cpp
@@ -628,7 +628,7 @@ dl_error_t DLInternalConvertNoHeader( dl_ctx_t     _Context,
 	}
 
 	Writer.SeekEnd();
-	*_pNeededSize = Writer.Tell();
+	*_pNeededSize = (unsigned int)Writer.Tell();
 
 	return err;
 }
